BenchmarkDataNLP.jl

BenchmarkDataNLP.jl

Introduction

Functions

Introduction

BenchmarkDataNLP.jl is designed to assist users in generating text corpus datasets with parameterized complexity. By utilizing methods such as context-free grammars and Chomsky trees, this package enables the creation of datasets tailored for training natural language processing (NLP) models, including large language models (LLMs). The primary objective is to offer a controlled environment where the complexity of the datasets can be adjusted, facilitating efficient training and evaluation of NLP methods within manageable timeframes and computational resources.

Functions

BenchmarkDataNLP.add_random_suffix — Method

add_random_suffix(input::AbstractString) -> String

Takes a string and returns it with a random 5-character suffix added.

Returns a string

Examples

julia> add_random_suffix("hello")
"helloabc12"

BenchmarkDataNLP.greet — Method

greet()

Prints a friendly greeting message to the console.

Returns nothing

BenchmarkDataNLP.repeat_string — Method

repeat_string(input::AbstractString, times::Int) -> String

Repeats a string or character a specified number of times.

Returns a string

Examples

julia> repeat_string("hi", 3)
"hihihi"